Segrated classes to separate scripts as well as added validation for the classes
